+++ src/github/TurboIssueEvent.java
@@ -1,4 +1,4 @@
-package service;
+package github;
 
 import java.util.Date;
