+++ src/github/update/LabelUpdateService.java
@@ -1,4 +1,4 @@
-package service.updateservice;
+package github.update;
 
 import static org.eclipse.egit.github.core.client.IGitHubConstants.SEGMENT_LABELS;
 
@@ -8,7 +8,7 @@ import org.eclipse.egit.github.core.IRepositoryIdProvider;
 import org.eclipse.egit.github.core.Label;
 import org.eclipse.egit.github.core.client.PagedRequest;
 
-import service.GitHubClientExtended;
+import github.GitHubClientExtended;
 
 import com.google.gson.reflect.TypeToken;
