+++ src/prefs/ConfigFileHandler.java
@@ -1,4 +1,4 @@
-package storage;
+package prefs;
 
 import com.google.gson.*;
 import org.apache.logging.log4j.LogManager;
@@ -7,7 +7,6 @@ import util.HTLog;
 
 import java.io.*;
 import java.lang.reflect.Type;
-import java.nio.channels.NonWritableChannelException;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
