+++ src/util/TickingTimer.java
@@ -1,4 +1,4 @@
-package service;
+package util;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
