+++ src/util/events/ModelUpdatedEvent.java
@@ -1,4 +1,4 @@
-package backend.assumed;
+package util.events;
 
 import backend.interfaces.IModel;
 import backend.resource.MultiModel;
